fix(config-ui): show jenkins scope tips condition error
