+++ config-ui/src/layouts/base/styled.ts
@@ -77,7 +77,6 @@ export const Sider = styled.div`
 `;
 
 export const Inner = styled.div`
-  position: relative;
   display: flex;
   flex-direction: column;
   flex: auto;
