+++ config-ui/src/pages/blueprint/detail/blueprint-detail.tsx
@@ -16,16 +16,18 @@
  *
  */
 
-import React, { useState } from 'react';
+import React, { useState, useMemo } from 'react';
 import type { TabId } from '@blueprintjs/core';
 import { Tabs, Tab } from '@blueprintjs/core';
 
 import { PageLoading } from '@/components';
+import { Plugins } from '@/plugins';
 
 import type { UseDetailProps } from './use-detail';
 import { useDetail } from './use-detail';
 import { Configuration } from './panel/configuration';
 import { Status } from './panel/status';
+import * as S from './styled';
 
 interface Props extends UseDetailProps {}
 
@@ -36,32 +38,46 @@ export const BlueprintDetail = ({ id }: Props) => {
     id,
   });
 
+  const showJenkinsTips = useMemo(() => {
+    const jenkins = blueprint && blueprint.settings?.connections.find((cs) => cs.plugin === Plugins.Jenkins);
+    return jenkins && !jenkins.scopes.length;
+  }, [blueprint]);
+
   if (loading || !blueprint) {
     return <PageLoading />;
   }
 
   return (
-    <Tabs selectedTabId={activeTab} onChange={(at) => setActiveTab(at)}>
-      <Tab
-        id="status"
-        title="Status"
-        panel={
-          <Status
-            blueprint={blueprint}
-            pipelines={pipelines}
-            pipelineId={pipelineId}
-            operating={operating}
-            onRun={onRun}
-            onUpdate={onUpdate}
-            onDelete={onDelete}
-          />
-        }
-      />
-      <Tab
-        id="configuration"
-        title="Configuration"
-        panel={<Configuration blueprint={blueprint} operating={operating} onUpdate={onUpdate} onRefresh={onRefresh} />}
-      />
-    </Tabs>
+    <S.Wrapper>
+      <Tabs selectedTabId={activeTab} onChange={(at) => setActiveTab(at)}>
+        <Tab
+          id="status"
+          title="Status"
+          panel={
+            <Status
+              blueprint={blueprint}
+              pipelines={pipelines}
+              pipelineId={pipelineId}
+              operating={operating}
+              onRun={onRun}
+              onUpdate={onUpdate}
+              onDelete={onDelete}
+            />
+          }
+        />
+        <Tab
+          id="configuration"
+          title="Configuration"
+          panel={
+            <Configuration blueprint={blueprint} operating={operating} onUpdate={onUpdate} onRefresh={onRefresh} />
+          }
+        />
+      </Tabs>
+      {showJenkinsTips && (
+        <S.JenkinsTips>
+          <p>Please add the "Jenkins jobs" to collect before this Blueprint can run again.</p>
+        </S.JenkinsTips>
+      )}
+    </S.Wrapper>
   );
 };
